# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Esto es la documentacion del sitio web del Super Mercado mediante la pagina de Swagger.  Puedes encontrar mas sobre   Swagger en [http://swagger.io](http://swagger.io) o entrar a [irc.freenode.net, #swagger](http://swagger.io/irc/).",
    "version": "1.0.0",
    "title": "Super Mercado",
    "contact": {
      "email": "alexbaptista.abf660@gmail.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "host": "localhost",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/compra/buscar/{Id}": {
      "get": {
        "tags": [
          "compra"
        ],
        "summary": "encontrar una compra por su ID",
        "description": "Retorna una compra mediante el ID",
        "operationId": "bucarcompra",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "ID de la compra que se desea encontrar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/compra"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/detalle_compra/buscar/{Id}": {
      "get": {
        "tags": [
          "detalle de compra"
        ],
        "summary": "encontrar un detalle de compra por su ID",
        "description": "Retorna un detalle de compra mediante el ID",
        "operationId": "bucarDetalleCompra",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "ID del detalle de compra que se desea encontrar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/detalleCompra"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "compra": {
      "type": "object",
      "required": [
        "id_cliente",
        "id_pago",
        "id_detalle_compra"
      ],
      "properties": {
        "id_cliente": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        },
        "id_pago": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        },
        "id_detalle_compra": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        },
        "fecha_caducidad": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        }
      }
    },
    "detalleCompra": {
      "type": "object",
      "required": [
        "id_producto",
        "cantidad",
        "precio"
      ],
      "properties": {
        "id_cliente": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        },
        "cantidad": {
          "type": "integer",
          "format": "int64",
          "example": "2"
        },
        "precio": {
          "type": "integer",
          "format": "int64",
          "example": "4"
        }
      }
    }
  },
...
}